fix(workspace): wire color-pipeline into squarebob bin + sweep pre-existing clippy lints
Two unrelated bundles of fixes landed together because the
release build broke as soon as `cargo check` started enforcing
strict clippy across the workspace.

## Bin-side dep

`crates/color-pipeline` was declared in `[workspace.dependencies]`
and pulled into `render-shared`, but the squarebob bin itself
imports `color_pipeline::*` from its color.rs and never listed
it under `[dependencies]`. Added the workspace-true line right
next to `egui-colorpicker`.

## Pre-existing clippy sweep

`bootstrap.py c` runs strict `clippy --workspace --all-targets
-D warnings`. Lint accretion in the codebase made it fail clean.
Mass-cleaned what was directly blocking the green build:

* `pt-mats` lib test — `{frac:.0%}` is not a valid Rust format
  spec; switched to named-positional `{pct:.0}% , pct = frac * 100.0`.
  This was a real compile error in the test target.
* `gpu-mem` — deleted two stale Windows-only tests that referenced
  `parse_reg_value`, the helper removed when the Windows query
  path collapsed onto DXGI.
* `render-shared::default_material_overrides` — replaced
  `let mut b = MO::default(); b.seed = ...;` with struct-literal
  `MO { seed: ..., ..MO::default() }` per
  `clippy::field_reassign_with_default`.
* `egui-colorpicker::hs_picker` / `v_slider` — collapsed
  `if cond { if let ... { body } }` into `if cond && let ...
  { body }` per `clippy::collapsible_if`.
* `pt-wavefront::pipeline` — dropped the `as u64` cast on a
  field already typed `u64`.
* `pt-megakernel::compute` — collapsed `if pathguide_enabled
  { if let Some(pg_bgs) = ... }` (one of five identical
  patterns; rest are still flagged and on the follow-up list).
* `pt-denoise-oidn` — same `if let && let` collapse.
* `playa-ae::attr_editor` — `PickerConfig { alpha_enabled:
  false, ..default() }` instead of mutate-after-default.

Release build verified clean (`python bootstrap.py b`, 52 s).
Strict clippy (`bootstrap.py c`) still has the four remaining
collapsibles in `pt-megakernel::compute` — they need a flow
restructure rather than a cosmetic rewrite, so they wait for a
dedicated cleanup pass.

+++ crates/pt-wavefront/src/wavefront/pipeline.rs
@@ -254,7 +254,7 @@ impl WavefrontPipeline {
         // Largest single binding — must fit under WebGPU's
         // `max_storage_buffer_binding_size` per buffer.
         let per_pixel_single = ray_sz.max(hit_sz).max(aov_sz).max(1);
-        let binding_limit = device.limits().max_storage_buffer_binding_size as u64;
+        let binding_limit = device.limits().max_storage_buffer_binding_size;
         let max_pixels_binding = (binding_limit / per_pixel_single).max(1);
         // Total per-frame VRAM consumed by the wavefront per-pixel
         // buffer set (2 ray + 1 hit + 2 AOV). The ping-pong ray
